+++ src/commands/add_traveler.rs
@@ -111,7 +111,7 @@ mod tests {
         let mut bot = TestBot::new(db, "/addtraveler");
         let err = CommandError::EmptyInput.translate_default();
         assert!(
-            bot.last_message()
+            bot.dispatch_and_last_message()
                 .await
                 .is_some_and(|msg| msg.starts_with(&err))
         );