+++ src/commands/list_transfers.rs
@@ -63,31 +63,30 @@ mod tests {
     use crate::{
         db::db,
         i18n::{self, Translate, translate_default, translate_with_args_default},
-        tests::TestBot,
+        tests::{TestBot, helpers},
         transfer::Transfer,
         traveler::Name,
     };
     use maplit::hashmap;
     use std::str::FromStr;
-    use teloxide::types::ChatId;
 
     test! { list_transfers_ok,
         let db = db().await;
         let mut bot = TestBot::new(db.clone(), "");
 
         // Add travelers "Alice", "Bob" and "Charlie"
-        add_traveler(&mut bot, "Alice").await;
-        add_traveler(&mut bot, "Bob").await;
-        add_traveler(&mut bot, "Charlie").await;
+        helpers::add_traveler(&mut bot, "Alice").await;
+        helpers::add_traveler(&mut bot, "Bob").await;
+        helpers::add_traveler(&mut bot, "Charlie").await;
 
         // Transfer 100 from Alice to Bob
-        transfer(&mut bot, "Alice", "Bob", 100.0).await;
+        helpers::transfer(&mut bot, "Alice", "Bob", 100.into()).await;
         // Transfer 50 from Bob to Charlie
-        transfer(&mut bot, "Bob", "Charlie", 50.0).await;
+        helpers::transfer(&mut bot, "Bob", "Charlie", 50.into()).await;
 
         // List transfers
         bot.update("/listtransfers");
-        let transfers = Transfer::transfers(db, ChatId(bot.chat_id())).await.unwrap();
+        let transfers = Transfer::transfers(db, bot.chat_id()).await.unwrap();
         // Check that two transfers has been recorded
         assert_eq!(transfers.len(), 2);
         let response = transfers
@@ -103,18 +102,18 @@ mod tests {
         let mut bot = TestBot::new(db.clone(), "");
 
         // Add travelers "Alice", "Bob" and "Charlie"
-        add_traveler(&mut bot, "Alice").await;
-        add_traveler(&mut bot, "Bob").await;
-        add_traveler(&mut bot, "Charlie").await;
+        helpers::add_traveler(&mut bot, "Alice").await;
+        helpers::add_traveler(&mut bot, "Bob").await;
+        helpers::add_traveler(&mut bot, "Charlie").await;
 
         // Transfer 100 from Alice to Bob
-        transfer(&mut bot, "Alice", "Bob", 100.0).await;
+        helpers::transfer(&mut bot, "Alice", "Bob", 100.into()).await;
         // Transfer 50 from Bob to Charlie
-        transfer(&mut bot, "Bob", "Charlie", 50.0).await;
+        helpers::transfer(&mut bot, "Bob", "Charlie", 50.into()).await;
 
         // List transfers related to Alice
         bot.update("/listtransfers Alice");
-        let transfers = Transfer::transfers_by_name(db.clone(), ChatId(bot.chat_id()), Name::from_str("Alice").unwrap()).await.unwrap();
+        let transfers = Transfer::transfers_by_name(db.clone(), bot.chat_id(), Name::from_str("Alice").unwrap()).await.unwrap();
         // Check that only one transfer is returned
         assert_eq!(transfers.len(), 1);
         let response = transfers
@@ -126,7 +125,7 @@ mod tests {
 
         // List transfers related to Bob
         bot.update("/listtransfers Bob");
-        let transfers = Transfer::transfers_by_name(db, ChatId(bot.chat_id()), Name::from_str("Bob").unwrap()).await.unwrap();
+        let transfers = Transfer::transfers_by_name(db, bot.chat_id(), Name::from_str("Bob").unwrap()).await.unwrap();
         // Check that two transfers are returned
         assert_eq!(transfers.len(), 2);
         let response = transfers
@@ -152,11 +151,11 @@ mod tests {
         let mut bot = TestBot::new(db.clone(), "");
 
         // Add travelers "Alice" and "Bob"
-        add_traveler(&mut bot, "Alice").await;
-        add_traveler(&mut bot, "Bob").await;
+        helpers::add_traveler(&mut bot, "Alice").await;
+        helpers::add_traveler(&mut bot, "Bob").await;
 
         // Transfer 100 from Alice to Bob
-        transfer(&mut bot, "Alice", "Bob", 100.0).await;
+        helpers::transfer(&mut bot, "Alice", "Bob", 100.into()).await;
 
         // List transfers related to Charlie -> not found
         bot.update("/listtransfers Charlie");
@@ -165,15 +164,5 @@ mod tests {
             &hashmap! {i18n::args::NAME.into() => "Charlie".into()},
         );
         bot.test_last_message(&response).await;
-    }
-
-    async fn add_traveler(bot: &mut TestBot, name: &str) {
-        bot.update(&format!("/addtraveler {name}"));
-        bot.dispatch().await;
-    }
-
-    async fn transfer(bot: &mut TestBot, sender: &str, receiver: &str, amount: f64) {
-        bot.update(&format!("/transfer {sender} {receiver} {amount}"));
-        bot.dispatch().await;
     }
 }