+++ src/commands/set_currency.rs
@@ -57,7 +57,6 @@ mod tests {
     };
     use maplit::hashmap;
     use std::str::FromStr;
-    use teloxide::types::ChatId;
 
     test! { set_currency_ok,
         let db = db().await;
@@ -95,7 +94,7 @@ mod tests {
 
         // Retrieve traveler "Alice" and their expenses
         let traveler =
-            Traveler::db_select_by_name(db.clone(), ChatId(bot.chat_id()), &Name::from_str("Alice").unwrap())
+            Traveler::db_select_by_name(db.clone(), bot.chat_id(), &Name::from_str("Alice").unwrap())
                 .await
                 .unwrap()
                 .unwrap();